Use MCDCMatrix in shaders
A few changes here. The motivation is to support using the MCDCMatrix
in the vertex shader where possible. This has the advantage of
performing the MCVC x VCDC multiplication in double precision
on the CPU as opposed to float precision on the shader. This fixes
a couple test failures.

Along with that change cleanup a bit how we determine what uniforms
to set. There were some uniforms being set when they did not exist
in the program. Provided a more general approach for passing that
type of information between the code that builds the shader program
and the code that sets the uniforms.

Add a substitute helper function that does the same job as replace
but returns a bool if a replacement was made. Also operates on
the source string by reference.

make a more generic way for texture units to be passed to a mapper.

Cleaned up some long lines to be shorter.

Change-Id: I7f8cbc4892e57cd4c3bed11d497940d1f2a7c63b