Add a subclass for faster composite dataset mapping
This class is a faster implementation of vtkCompositePolyDataMapper
that requires its inputs to adhear to some constraints. The big change
is that it required adding functions to append to a VBO or IBO.

Part of this change is I decoupled the opacity from scalar coloring.
Previously a change in opacity required rebuilding the VBO if
scalar coloring was on. While this is still the behavior, it can
now be optimized to not do that.  It doesn't made sense for
the property's opacity to be mixed in with scalar colors, it
was just done that way because that we all the old pipeline
could do.

Also cleaned up a few long lines while I was a it.

Change-Id: I3a06b502f154c4da7465e53eeeaeab9492bca88f