Wait for resize notification in SetSize(...)
Calling XSync(...) is not enough. This causes a timing issue
went SetSize(...) is called following be a Render(...). In
Render(...) the window is queried for the window size which
will not have be resize yet so we get the old size! ( seen
particularly in vtkWeb ). Add loop to wait for ConfigNotify
event with the new size.

Change-Id: I72d34cae16215b6887467496df23d7a8e0bdcd37