Fix conic point identification in the path renderer.
The numeric value of FT_CURVE_TAG_CONIC is 0, so the translation
to local ControlType failed when testing

  if (fttag & FT_CURVE_TAG_CONIC) { ... }

Replaced the translation with a more robust approach.

This wasn't readily apparent using the fonts embedded in VTK, as those
fonts consist entirely of linear and cubic curves.

Also fixed this functiont to use the pen_[xy] variables, which account
for kerning.

Change-Id: Ibcd7d28365ecad22d62073dc8bbf89405478f791