vtkModuleTop: handle UsrMove-like setups
The problem is that when the package is found via a directory with a
symlink which points into the install prefix, we should use the
configured install prefix. Example:

    /usr (install prefix)
    /usr/lib/cmake/vtk-6.2 (configuration's path)
    /lib -> /usr/lib (UsrMove)
    /lib/cmake/vtk-6.2 (path VTK was found by)

The problem here is that when we traverse up the 3 components to try and
find '/usr', we end up finding '/' instead. To fix this, we detect this
by resolving the found path and the expected path to remove all
symlinks. If these are the same, use the configured path, otherwise go
up the tree the correct number of components.

Still unsupported is finding VTK by a path with a symlink into the
prefix and VTK has been relocated since being installed.

Fixes #15065.

Change-Id: I181a126382941eb7e33304d426719abd5769048b