Fix arg conversion of const ref arg via constructor.
There was a simple mistake in the code that builds the "ArgCheck"
string that the wrappers use for argument type checking.  As a
result, the "argument check" was not distinguishing between const
reference args and non-const reference args.  A "const reference"
arg type can satisfied via implicit conversion from another type,
while a non-const reference arg cannot.  The new test specifically
tests for this.

Change-Id: I59725536496dd6d80b64ace09403161ccd89c9b4