Fix lights issue in external render environments
With this change, the external VTK renderer disables automatic light
creation. In each frame, OpenGL lights are queried from the OpenGL
state and a new headlight is added in VTK for each enabled OpenGL
light. This way VTK objects in the scene are lit with lights from the
external OpenGL application.

NOTE: VTK lights created in the pipeline are discarded and only
the externally created OpenGL lights are regarded.