Kill TDX timer when using no WindowHandle
From KillTimer(HWND hWnd, UINT_PTR uIDEvent) documentation:
uIDEvent [in]: The timer to be destroyed. If the window handle passed to
SetTimer is valid, this parameter must be the same as the nIDEvent value
passed to SetTimer. If the application calls SetTimer with hWnd set to
NULL, this parameter must be the timer identifier returned by SetTimer.

Fixes #15334

Change-Id: Ifd88c790f4955dcd352004e2f5cba22a900e02de