Visual Studio 2013 requires the <algorithm> header.
This is required for std::max() and std::min().

I think (but don't quote me on this) that the older MSVC
compilers used to automatically include a MSVC specific
header called xutility where these were defined.
With respect to other comilers e.g. gcc, it is possible
that one of the other includes is adding the algorithm header in.

Change-Id: I9a11e00bed8887f421cbde317e3f2ead29e8f39a
Signed-off-by: Andrew Maclean <andrew.amaclean@gmail.com>