Use the appropriate components for transfer functions
When the flag IndependentComponents is ON, pass each component
through the transfer function.
When it is OFF and there are two components, the first component
should be used for color function and the second for opacity
function.
When it is OFF and there are four components, only the last
component should be passed through opacity transfer function.