Fix how we import our argparse fallback implementation.
In environments where Python 2.6 is still in use, there is no
argparse module included by default.  For these situations, we
have included our own copy, called _argparse.py, which we usually
import in the except block in case the default is not present.  At
some point, the vtk.web package included this file, but at some
point it was removed and because few sites still use Python 2.6,
it was not caught.

This fix puts the _argparse.py in a single location (except for
where it is needed at build time).  This location is vtk.util,
and everything now looks there in the fallback case.

Change-Id: Id64241ffbe4e3cec7e96dae386da1a6385c5f6ed