Improve appearance of small-sized mathtext.
The alpha mask returned by Matplotlib for mathtext renderings
appears very dim when font size is small.

This patch uses the same trick in MatplotlibMathTextUtilities
as FreeTypeTools, which doesn't blend RGB values when the
background alpha == 0. This gives a bolder appearance to the
text, like it was before the text background color options
were added.

Also clamped the tprop opacity setters to [0., 1.] and updated
a baseline that had small mathtext.