Fixed compiler error with div() introduced with Wconversion fixes
\Filters\Statistics\vtkOrderStatistics.cxx(346): error C2440: 'initializing' : cannot convert from 'lldiv_t' to 'div_t'

Interesting case.  Due to changing one of the params to div()
from int to vtkIdType (which can be long, long long, int, etc.)
the templated div() changed to one that returned the lldiv_t
struct instead of the div_t structure.

Switched to ldiv() because there seems to be only one version of it.