Remove tcl [info command "rtExMath"] check from test.
Dunno what this does, but without it the baseline doesn't
match the rendering. The image difference was small enough
that this test has been passing for years, until the slight
text changes pushed it over the tolerance.

The link below suggests that it was something to check to see
if the script was running inside of VTK's TCL interpreter?

http://stackoverflow.com/questions/18457569/info-command-rtexmath-in-tk-with-vtk

In any case, the outline actor is definitely there in the
baseline, so let's just add it unconditionally.