Remove warning when a array name for labeling is not found.
When interactively selecting data with multiple sources
(select the cell/point under the mouse pointer)
you want to see a label for the source where the field exists
and not see it if the field does not exit.

Not seeing the label is enough indication that the array does not
exist.