Add support for picking indirection arrays
At the same time this patch completely changes how cell data is
handled. In the past we would duplicate any shared vertices so
that we could apply the cell data uniquely to each vertex.
That causes a huge increase in size for the data as a typical
mesh has each vertex shared by 6 triangles.

The new approach we create a texture buffer where we store the
cell data. Then in the fragment shader we use the gl_PrimitiveID
to index into the texture buffer to lookup the relevent cell data.
With this change gl_PrimitiveID support is now required not just
for picking but for everything.