Convert Cocoa to default to OpenGL 3.2
OSX is a bit more strict on its opengl so a bunch more changes were
required such as using in/out instead of attribute/varying. Also
handle gl_FragData properly for 3.2 as an output variable.

glew had a bug preventing glGenVertextAttribute... from being
defined. Made a change to correct that.