Various fixes for vtkPolyLine{Source,Representation}
Added vtkPolyLineSource::Resize() method to support memory-preserving
changes in the number of points in the poly line
source. vtkPolyLineRepresentation uses this Resize() method when the
number of handles is changed. Added initialization of new handle
positions in vtkPolyLineRepresentation when handles are added to fix
uninitialized value errors reported by valgrind. Modified
vtkPolyLineRepresentationTest1 to check that handle positions are
preserved when number of handles increases.