ENH: Added edge annotations to vtkCornerAnnotation
The vtkCornerAnnotation class provided an easy means for overlaying
2D text annotations onto the corners of a rendering.  Medical
imaging applications typically use S/I A/P R/L annotations in the
center of the edges of a visualization to denote patient orientation,
but this was not yet supported.  This enhancement adds support for
edge annotation overlays, allowing any combination of edge and corner
annotations, with no change to existing code that uses the corner
annotation functionality.  The example has been updated to test both
functions.  It may be worth considering renaming the class to
vtkEdgeAnnotation to more generally characterize its functionality
(as corners are parts of edges, but not vice versa).