Fix a few issues exposed by PV testing
A couple fixes.  The big one is fixes to picking for
CompositePolydataMapper2 but there are a couple other minor
fixes in there as well.

SetPixelData was not turning off depth tests.

Some cases with a lot of degenerate cells could result
in negative array sizes.

Testing for a stereo capable window on Win32 was not correct