ENH: Modified vtkCornerAnnotation testing to improve code coverage
Modified TestCornerAnnotation to test a larger subset of
vtkCornerAnnotation's behavior.  Switched from showing annotations
over polygonal geometry to over a image, allowing for testing of
the window/level and slice display functionality.  As a future
enhancement, may want to try annotating over a 3D image to fully
test the slice number substitutions.  Remaining uncovered code is
primarily in the readback/copy/clear routines and some of the text
sizing. Had to add vtkImageSources to the module.cmake/TEST_DEPENDS
for Rendering/Annotation in order to use a canned image source -
this could be avoided by generating an image from scratch.