BUG #15438: Handle NaN when computing array range.
A small tweak to the vtkDataArrayPrivate makes it possible for us to
handle NaN in arrays when computing ranges. We should not include NaN
when computing the range. By ensuring that the "left" argument to the
min/max functions is never a NaN, we are assured that they min/max calls
won't return a NaN.

Also added a test for the same.