Expose mode-shape-range meta-data to downstream filters.
Filters downstream from vtkExodusIIReader has access to mode_shape in
FieldData, but there was no way of getting information for the range of
the mode shapes. This changes added support for a new "mode_shape_range"
array that provides this meta-data.

Also adding "mode_shape" and "mode_shape_range" arrays only when
HasModeShapes is true. Otherwise, these variables are meaningless.