Fix blank slices that appear in some TIFF volumes.
The logic for skipping subfiles that aren't part of the primary image
(e.g. subfiles that contain reduced-resolution slices) was incorrect.
When the reader encountered such subfiles, it would simply move on to
the next slice of the image, without setting the voxel values for the
current slice.  Now, the "slice" counter and the "page" counter are
distinct from one another, since not all pages become image slices.