Fix vtkAxis scaling when TileScale is set on the renwin.
Now vtkAxis will show the same number of ticks and the
same relative text size when magnifying the image. Text
will still be a little funky if the scaling is anisotropic,
but at the moment our freetype engine doesn't support
anisotropic dpi (not sure if the MPL backend can or not).

New API: vtkContextScene::GetLogicalTileScale() returns
the renderwindow tilescale or (1, 1) depending on whether
ScaleTiles is enabled.

There are some random fixes where vtkAxis objects were being
used without a scene, which led to segfaults when trying to
look up the tile scale internally.