BUG #15155: Fix issues with temporal interpolator.
vtkTemporalInterpolator changed the output dataobject to be same ptr as
the internally cached dataobject. That meant that when the output was
reset for a subsequent execution, the cached dataobject would get
reset/reinitialized too! That was causing the report bug. Fixed by not
changing the output data object ptr, instead shallow copying the cached
value to it (as is the norm in VTK filters).