ENH: CanReadFile does not fail for newer file version.
CanReadFile does not fail for newer file versions to enable clients (ParaView)
to distinguish between reader failures caused by a mismatch between the file
and the reader, and reader failures that cannot be fixed by changing the reader.

A warning will be printed when reading a file with a newer version than the reader.

Legacy readers already behave this way.