Restore concept of InteractiveUpdateRate for vtkSmartVolumeMapper
vtkSmartVolumeMapper decides whether its internal mappers should
auto-adjust or use a specific sample distance based on whether
the render is interactive or still. The InteractiveUpdateRate
member variable helps decide this.
This variable was removed in dcd2ee8 when the sample distance API
was extended to vtkSmartVolumeMapper.
This change just affects the vtkSmartVolumeMapper available under
the OpenGL2 backend.