Added unary plus and fixed scientific notation bug in vtkFunctionParser
This fix is in reference to bug report 0015557. Originally, the expression
'3.0e+01' was being evaluated as '3.0e1 + 1 = 31'. It now behaves correctly.
Additionally, a unary plus operator has been added to the general case
(previously, the inclusion of a unary plus resulted in an error). The test
program for vtkFunctionParser has been updated to include unary plus tests
and a scientific notation test.