Correct and augment the behavior of vtkPolygonBuilder
This fix is in reference to bug report 0015582. vtkPolygonBuilder manifested
two problems: triangles that were not adjacent to any triangles that were
previously added to the builder were discarded, and polygons with internal
vertices resulted in overlapping cells (the class was explicitly not designed
to handle polygons with internal vertices). The new version fixes these two
issues, and test TestCutter has been added to check these issues as well.