Extend vtkPolygonBuilder to return multiple disjoint polygons, if present.
This fix is in reference to bug report 0015582. In cases where a single cell
is cut into multiple cells, it is necessary to return disjoint polygon elements
in vtkPolygonBuilder. Additionally, a test provided by Andreas Buykx has been
included that tests the ability of vtkPolygonBuilder to construct multiple
disjoint polygons.