BUG #15662: vtkAppendPolyData does not remove UPDATE_EXTENT
For a pipeline as the following:

  // image0 image1 image2
  // dmcubes0 dmcubes1 dmcubes2
  // \          |       /
  // vtkAppendPolyData
  // mapper/actor
  // renderer

dmcubes is a discrete marching cubes alg.
The pipeline forwards UPDATE_EXTENT for all its inputs.
This results in the UPDATE_EXTENT from the first image being compared with the WHOLE_EXTENT of the
second image which results in a mismatch.

We fix this by letting downstream request a subset of connection 0, for connections >= 1
we send their WHOLE_EXTENT as UPDATE_EXTENT.