Ensure that entries in hierarchy file are unique.
Previously, it only checked whether the entire line in the file was
unique.  Instead, it should verify that the entry itself is unique.
This became an issue in some C headers that had declarations of the
form "typedef struct X {} X;", causing X to appear both as a typedef
and as a class type.