Fix classes that broke the "rule of three"
The expanded python wrapping identified a few classes that broke
the rule of three: they defined a custom destructor (indicating that
they performed some kind of resource management), but did not define
a custom copy constructor or a custom assignment constructor.  Some
builds were generating a warning because of the missing methods.