BUG #15746: Fixes issues with Merge Points in vtkAppendFilter.
vtkAppendFilter was incorrectly setting the number of tuples in output
attribute arrays when "Merge Points" was on. It simply accumulated the
number of points in each of the inputs, which is wrong when the datasets
had points that get merged. This was resulting in the output array having
uninitialized elements. Fixed that.