A fix for picking combined with depth peeling
A fix for picking combined with depth peeling and some cleanup.
Picking (not using a hardware picker) was not clearing the
LastRenderingUsedDepth peelign flag and that was causing issues.
Switched to that that flag is not sued but instead the information
objects are used.

(cherry picked from commit 7e23fd65333d8c0f44b7b0770bf95cc5cfdaa04a)