Accelerated vtkSpanSpace with parallel sort
A vtkSMPTools::Sort() method was added (which is a drop in replacement
for std::sort()). Under the hood, in TBB tbb::parallel_sort() is used.
Other SMP implementations (Sequential,Simple,Kaapi,OpenMP) are for now
using std::sort().

This new vtkSMPTools capability was then used to accelerate the
sort process in vtkSpanSpace as well as vtkCheckerboardSplatter.