Prevent DataSetSurfaceFilter to merge points of quadratic meshes.
The filter internally call a vtkUnstructuredGridSurfaceFilter to
handle subdivision of quadratic cells. By default this filter
merge points in the generated mesh. This prevent to perform
a correct visualization of non-continuous fields defined
with duplicated points (eg. crack or fault).

This patch turn the merging feature off for such meshes.