Add android vr example and fixes to support it
This topic adds a volume rendering example to android builds.  It also
fixes a number of issues to make it work with android including on
Samsung devices which have a different OpenGL driver.

VR use of 1D textures have been replaced with 2D textures as
1D textures are not supported on OpenGL ES.

Some minor build inprovements are included as well.