Defer updates until an update request is received.
Rerunning the reader when a Set[IVar] method is called is not correct.
These should just updated the internal state and mark the filter/reader
as modified and rerun when needed by the pipeline.

Also removed TimeSteps ivar, which was just a temporary array needed in
a single function scope and simplified the timestep/info object
initialization.