BUG #15763: Avoid modifying camera when it hasn't changed.
Fixes BUG #15763. The bug was caused because different ranks in ParaView
made mismatching decision about whether to the HardwareSelector needed
to re-render for the selection. This bug was causing the server side
camera to always be modified since it would always create and set new
vtkMatrix4x4 instances whenever the cameras were synchronized between
ranks. Fixed not doing that and instead using API on vtkMatrix/vtkCamera
that calls Modified() only when things truly change.