BUG #15632: Fix exodus performance issue with spatial files.
When files in an exodus file series were spatial partitions, reading
timesteps from each one of them is not only wasteful but also slows down
the reader considerably. In this fix, we update vtkPExodusIIReader to
pass the timesteps information from the first reader to all the rest and
avoid the other filters from re-reading the time information. Since the
reader has to skip through files to read time information, it can have a
serious performance impact. This fix addresses the issue with minimal
refactoring of the reader.