BUG 15797: Fix text property ID computation
Only a few settings defined by the vtkTextProperty were used to
compute the text property ID in MapTextPropertyToId(). This can cause
incorrect rendering when settings that weren't considered, such as the
shadow setting, were changed.

Added a more general hashing function, HashBuffer(), to accommodate
all the vtkTextProperty settings, and use this to compute the
vtkTextProperty hash. HashString() is no longer used, but remains for
backwards compatibility.

Added a test to ensure that changing a setting results in the
computation of a different ID.