Improve the shadow maps in VTK
This patch changes how the shadow maps work so that
less shadow code is in the mappers. Now the mappers get
a key that contains the shadow map pass and they let
that pass handle setting up shader code and uniforms.

Also changed the shadow map algorithm slightly.  The old approach did
a 1/3 weighting of pixels (0,0) (1,0) and (0,1)  This new algorithm
does a 2x2 box convolution against the 3x3 pixel neighborhood.
It takes 9 tex lookups but produces much nicer results.