Added option to display range labels
Off by default, this new option adds labels at the ends of the axis
with the range values. A separate printf-style format string can be
specified for these labels. If a tick label overlaps one of the range
labels, it will not be displayed.

Modified TestAxes and its baseline to exercise this new option.