BUG 12531: Fixed problem with AutoOrient
When resizing a vtkScalarBarWidget while AutoOrient is enabled, the
scalar bar orientation may change while the dimension whose size is
changing will not. This is odd behavior that can result in the scalar
bar jumping around during resizing or the widget completely
collapsing, making it impossible to interact with the widget.

Fixed this by disallowing orientation swaps when the scalar bar
representation is not moving.

Modified TestScalarBarWidget.cxx to test this change.