Fix rendering gradient background with vtkRenderLargeImage
This commit fixes incorrect rendering of a gradient background in images
generated by vtkRenderLargeImage. Previously, the gradient background was
rendered in full on each tile. Now, new background colors for each tile are
calculated by linearly interpolating between the colors specified for the full
window.

A Python test, similar to the existing 'largeImageOffset' and
'largeImageParallel' tests, is added.

Fixes #15715