Added new class vtkStaticPointLocator
A new class for constructing a point locator. The class is threaded
and templated for performance. It is a static locator, meaning that
it can only be built once (and incremental insertion is not allowed).
The code also specially treats 32-bit versus 64-bit id types for
performance and to reduce memory consumption.

Additional documentation and code enhancements were made to related
classes. A minor bug in vtkPointLocator related to not using the
right distance measure (dist2 vs maxDist2) was also corrected.