ENH: Unit tests for MaskPoints and PMaskPoints
MaskPoints test coverage was less than 30%. This patch covers 97% of
MaskPoints and 100% of PMaskPoints.

This patch provides a unit test that covers nearly 100% of the
code. Some of the code in MaskPoints is only exercised when run with
multiple processors. This required a unit test for PMaskPoints which
uses MPI. PMaskPoints was not present in any tests.