ThirdParty: Teach update script to store upstream hash in commit message
Store the upstream commit sha1 in the upstream snapshot commit message
subject line.  Format the subject line to make it easy to find the most
recent snapshot commit in history to use as the base for the next
update.

In order to help transition existing scripts, take the basehash from
them for now and allow an empty update commit.