Remove AGL.framework from OS X OPENGL_LIBRARIES.
The AGL framework for Carbon is deprecated by VTK and by OS X itself.
Linking to it causes forward compatibility problems because Apple has
removed it from the Xcode 10.10 SDK. This framework is added to the
OPENGL_LIBRARIES by FindOpenGL.cmake for CMake 3.4 and earlier, and
must be removed by vtkOpenGL.cmake.