Fixed interpolation for precise hits
When interpolating from input points to the output
volume; if any volume point was precisely on top of
an input point the interpolation failed. This was
corrected, some obsolete code was removed, and the
documentation was cleaned up. Also introduced the
notion of the power parameter P which will be fleshed
out in the next commit.