Combine vtk_wrap_python and vtk_wrap_python3.
The vtk_wrap_python macro was a copy-paste of vtk_wrap_python3, and this
commit reformulates it so that it calls vtk_wrap_python3 instead. Removing
vtk_wrap_python3 is not possible because it is the only way to wrap code
that does not belong to a VTK module, and it is used by external packages.