BUG 15914: Fix corrupt image from vtkImageResliceMapper
Ever since commit 263583d1, the IBO used by this mapper was no longer
being updated when the geometry changed.  For example, when the oblique
slice shape changed from a quadrilateral to a hexagon or vice-versa, the
rendered image became corrupted.  Adding a Modified() call to the cell
array fixed the issue.