Avoid duplication in vtkXMLDataSetWriter and vtkXMLCompositeDataWriter.
vtkXMLCompositeDataWriter, instead of duplicating code from
vtkXMLDataSetWriter to create the appropriate type of reader, now simply
call a new method on vtkXMLDataSetWriter to create the writer for
a specific dataset type.

Also avoid hardcoding filename extensions, instead using API on the appropriate
writer for the dataset type to get the extension string.