Hack HDF5 to build under ASan & UBSan
Parts of HDF are built as part of VTK’s build process.
It crashes and burns under ASan & UBSan thus preventing all
of VTK from being built & tested with these invaluable tools.
Filed as HDF bug HDFFV-8147 in 2012, but the HDF folks don't
seem willing or able to fix it.