Fix vtkTestingGL2PS module to work in Testing/External build
When this module is used by tests in the Testing/External build the
`VTK_CMAKE_DIR` value may point to a VTK install tree instead of the
source tree.  The vtkTestingGL2PS module and its RasterizePostScript
helper are not installed and so do not exist in the install tree.
Avoid using `VTK_CMAKE_DIR` to reference one module from the other.