Fix vtkVolumeRayCastSpaceLeapingImageFilter split.
This filter seems to have some bugs related to the increments it used
to march through the data, which make it break when multithreaded.  A
couple egregrious problems were fixed, but more seem to remain.  If the
volume is only split along the Z axis, the bugs don't seem to manifest.