Add modifier to translate center of the cylinder along the cylinder axis.
A new interaction state "TranslatingCenter" is added in order
to offer the possibility to translate the center of the axis along
the current axis. Use case: set the cylinder so it passes at one
position, move the center to it, then change the rotation axis
to set the second axis 'end-point'.
To use this feature, the user must select the center 'sphere' representation
then move the move while pressing CTRL key.