BUG: OpenGLPolyDataMapper2D regression
OpenGLPolyDataMapper2D for OpenGL was clipping the z components
between -1,0 or 0,1. The logic is a bit different in OpenGL2 versus
OpenGL. I believe it has somethiong to do with where the
transformations are performed. I'm pretty sure z-clipping should not
apply for 2D mappers. A new test was added that has polydata with a
large z range.

NOTE: To see the bug, download this patch and revert to the
original mapper:

git checkout master Rendering/OpenGL2

and
ctest -R TestLegendBoxActor2