Fix issues with volume rendering for dependent component data
When considering the components of the volume as dependent, the
last component (#1 for 2 component data and #3 for 4 component
data) is passed through the first opacity transfer function, set
with volumeProperty->SetScalarOpacity(0, piecewiseFunction) or
volumeProperty->SetScalarOpacity(piecewiseFunction).

Along the same lines, the gradients are computed based on the
last component scalar values for dependent component data. The
opacity is modulated based on these gradients by passing them
through the first gradient opacity transfer function.