Allow variables to be initialized with constructor args.
In vtkParse.y, there was no rule for variable initialization via
constructor args, for example "someclass variablename(arglist)".
Unfortunately, this syntax looks similar to function declaration
"someclass functionname(parameterlist)" and leads to ambiguity
because the parser cannot always distinguish between "arglist"
(a list of values) and "parameterlist" (a list of types).

The reason the parser has difficulty with this is that it lacks
the ability to tell which identifiers refer to types, and which
refer to values.  For now the parser demands that "arglist" begins
with a literal, since a literal is always a value.