fix ComputeOffAxisProjectionFrustum eyeseparation and near and far planes.
I have modified the far plane to scale with the distance to the object as well as the depth of the object. I have also modified the near plane to scale with the distance to the object. These changes when combined provide much better stereo support across a wide variety of object scales.