RK45 was setting delTActual wrong. Fixed.
RK45::ComputeNextStep was setting delTActual to delT before returning.
In one case, this was wrong. Since all code paths already set delTActual
correctly, removing the assignment fixed the problem.
For reference, the case where this produced the wrong result was
when the while loop ended with estError > maxError, in which case
the last value assignment of delT should not have been reflected
in delTActual.