Misc. fixes to vtkArrayListTemplate.
Contrary to the documentation, vtkArrayListTemplate::AddArrayPair() was
adding excluded array to the list. Fixed that to not add excluded
arrays.

Also made AddArrayPair() return a pointer to the newly creted array, if
any. This makes it possible for caller code to add the created array to
and output DataSetAttributes instance, for example.