Add vtkDualDepthPeelingPass.
This implements a new depth peeling algorithm that peels layers of
fragments from both the front and back of the translucent geometry at
once.

This also introduces vtkOpenGLRenderPass, which provides a virtual API
that allows custom shader code replacements without the need to edit
all mappers.

The older vtkDepthPeelingPass has been updated to use the new
vtkOpenGLRenderPass API, and is left in VTK as a fallback for drivers
that lack support for the dual peeling features (floating point textures
in particular).