vtkObjectFactory: add hook where extra init code can be added
When a class is already overridden (such as with vtkRenderWindow)
and you want to override it again you have to disable the first
override.  You can do this in the factory init code. This change
makes it so that you can do so without having to have your own copy
of the vtkObjectFactory.cxx.in etc.