take another pass at the LIC mapper
Try to clean up a few things I did not have time to
address in the first conversion. It now doesn't do the
wierd thing with setting its Output so that it can map
vectors to tcoords.  This is now properly passed as an
attribute.

There were a couple bugs realted to parallel composite LIC
one being a vertex coordinate system issues the other being
a MPI deadlock with icet