Fix the way that MINC IO handles null-termination.
When vtkMINCImageReader creates a vtkCharArray to store a NetCDF
attibute, it allocates one more byte than the number of chars, in
order to make room for a null terminator.  The way it did so was
unsafe and incompatible with the new vtkGenericDataArray changes.
This new code does the same thing, but in a compatible manner.