Clear to a transparent white background in RenderToImage mode
This allows the background values outside the bounds of the
volume to be white and transparent in the color image and white
in the depth image. This combination makes sure that the depth
image has an opaque white background outside the bounds of the
volume; while the color image has a transparent white background.