vtkFFMPEGWriter: Fix build with libavcodec57
Version 55 of the libavcodec API added an AV_ prefix to the PIX_FMT_
constants but retained a compatibility definition of the old constant.
Version 57 of the libavcodec API dropped the compatibility definition.
Select a constant name based on the API version in use.

Version 55.28.1 of the libavcodec API deprecated avcodec_alloc_frame in
favor of av_frame_alloc.  Version 57 dropped the old name.  Select a
function name based on the API version in use.