break core's dependence on ospray and begin to flesh out standalone
Using an interface class so that Volume module has no dependence on
OSPRay module. SmartVolume doesn't need to make any OSPRay calls since
the subclass takes over. Exposed it to OpenGL2 SmartVolumeMapper too.
Cloned the test that exercises this into the OSPRayModule.

The sublass is now the basis for a standalone OSPRayVolumeMapper
which will use OSPRay ONLY for the volume. Started a test for that.

warnings:
1) cloned smart volume still test hangs on my machine
2) the new test demonstrates that the volume is not yet composited with
surface geometry correctly.
4) the interface class needs to be reparented onto AbstractVolumeMapper.

/